Moves files to helpers
