+++ app/documentsAdapters/local/services/updateDocument.ts
@@ -1,5 +1,5 @@
 import fse from 'fs-extra';
-import findOrCreateDocuments from './findOrCreateDocuments';
+import findOrCreateDocuments from '../helpers/findOrCreateDocuments';
 import find from 'lodash/find.js';
 import extend from 'lodash/extend.js';
